+++ part5/bloglist-frontend/src/App.jsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from 'react'
 import Blog from './components/Blog'
+import LoginForm from './components/LoginForm'
 import blogService from './services/blogs'
 import loginService from './services/login'
 
@@ -28,58 +29,11 @@ const App = () => {
     }
   }, [])
 
-  const handleLogin = async (event) => {
-    event.preventDefault()
-    
-    try {
-      const user = await loginService.login({
-        username, password,
-      })
-      setUser(user)
-      blogService.setToken(user.token)
-      window.localStorage.setItem(
-        'user', JSON.stringify(user)
-      ) 
-      setUsername('')
-      setPassword('')
-    } catch (exception) {
-      setErrorMessage('Wrong credentials')
-      setTimeout(() => {
-        setErrorMessage(null)
-      }, 5000)
-    }
-  }
-
   const handleLogOut = () => {
     window.localStorage.removeItem('user')
     setUser(null)
     blogService.setToken(null)
   }
-
-  const loginForm = () => (
-    <form onSubmit={handleLogin}>
-      <h2>login to application</h2>
-      <div>
-        username
-          <input
-          type="text"
-          value={username}
-          name="Username"
-          onChange={({ target }) => setUsername(target.value)}
-        />
-      </div>
-      <div>
-        password
-          <input
-          type="password"
-          value={password}
-          name="Password"
-          onChange={({ target }) => setPassword(target.value)}
-        />
-      </div>
-      <button type="submit">login</button>
-    </form>      
-  )
 
   const blogsList = () => (
     <>
@@ -95,7 +49,7 @@ const App = () => {
   return (
     <div>
       {user === null ?
-        loginForm() :
+        <LoginForm setUser={setUser} setErrorMessage={setErrorMessage}/> :
         blogsList()
       }
     </div>
